requirements.txt
Django
psycopg2
djangorestframework
markdown
django-filter
pygments
dj-rest-auth
django-allauth
djangorestframework_simplejwt
yfinance
psycopg2-binary
drf-yasg